add story for form controls with native form element
